+++ CHANGELOG.md
@@ -15,6 +15,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/) 
 - [MINOR] Added `calculate_diff_days` and `calculate_diff_years` to `date_util`
 - [MINOR] Added `datetime_to_utc` to `date_util`
 - [MINOR] Added `init_external_loggers` to `logging`
+- [MINOR] Added `end_of_day` to `date_util`
 
 ### Changed
 - [MINOR] Updated `Requester` to send data correctly for PUT requests